feat(4.8): add and use abstract view
