+++ src/views/empty.js
@@ -1,30 +1,17 @@
-import { createElement } from '../utils/dom';
 import { Messages, Filters } from '../constants';
+import AbstractView from './abstract-view';
 
 const createEmptyTemplate = (activeFilter = Filters.Everything) => (
   `<p class="trip-events__msg">${Messages[activeFilter]}</p>`
 );
 
-export default class Empty {
-  #element = null;
-
+export default class Empty extends AbstractView {
   constructor(activeFilter) {
+    super();
     this.activeFilter = activeFilter;
-  }
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
   }
 
   get template() {
     return createEmptyTemplate(this.activeFilter);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
