+++ src/views/loading.js
@@ -1,23 +1,9 @@
-import { createElement } from '../utils/dom';
+import AbstractView from './abstract-view';
 
 const createLoadingTemplate = () => '<p class="trip-events__msg">Loading...</p>';
 
-export default class Loading {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class Loading extends AbstractView {
   get template() {
     return createLoadingTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
