+++ src/views/points.js
@@ -1,23 +1,9 @@
-import { createElement } from '../utils/dom';
+import AbstractView from './abstract-view';
 
 const createPointsTemplate = () => '<ul class="trip-events__list"></ul>';
 
-export default class Points {
-  #element = null;
-
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
+export default class Points extends AbstractView {
   get template() {
     return createPointsTemplate();
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
