+++ src/views/route.js
@@ -1,5 +1,5 @@
 import { formatTripDuration } from '../utils/date';
-import { createElement } from '../utils/dom';
+import AbstractView from './abstract-view';
 
 const formatRoute = (cities) => (
   cities.length < 4
@@ -15,28 +15,15 @@ const createRouteTemplate = (destinations, dateFrom, dateTo) => (
   </div>`
 );
 
-export default class Route {
-  #element = null;
-
+export default class Route extends AbstractView {
   constructor(destinations, dateFrom, dateTo) {
+    super();
     this.destinations = destinations;
     this.dateFrom = dateFrom;
     this.dateTo = dateTo;
   }
 
-  get element() {
-    if (!this.#element) {
-      this.#element = createElement(this.template);
-    }
-
-    return this.#element;
-  }
-
   get template() {
     return createRouteTemplate(this.destinations, this.dateFrom, this.dateTo);
-  }
-
-  removeElement() {
-    this.#element = null;
   }
 }
